+++ packages/components/src/Disclosure/Disclosure.css.d.ts
@@ -3,6 +3,7 @@ declare const styles: {
   readonly "details": string;
   readonly "summary": string;
   readonly "summaryWrap": string;
+  readonly "customSummaryWrap": string;
   readonly "content": string;
   readonly "openAnimation": string;
 };