+++ packages/components/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.29.0](https://github.com/GetJobber/atlantis/compare/@jobber/components@5.28.1...@jobber/components@5.29.0) (2024-08-29)
+
+
+### Features
+
+* **components:** Separate Chip from Chips Again ([#1995](https://github.com/GetJobber/atlantis/issues/1995)) ([1be6f9e](https://github.com/GetJobber/atlantis/commit/1be6f9e82dc856cfa1676a872832569e7d934afc))
+
+
+
+
+
+## [5.28.1](https://github.com/GetJobber/atlantis/compare/@jobber/components@5.28.0...@jobber/components@5.28.1) (2024-08-29)
+
+
+### Bug Fixes
+
+* **components:**  Update description of title prop in Disclosure component ([#1998](https://github.com/GetJobber/atlantis/issues/1998)) ([53196d6](https://github.com/GetJobber/atlantis/commit/53196d62aa0fbf83127e3bce2326bed308bf36da))
+
+
+
+
+
+# [5.28.0](https://github.com/GetJobber/atlantis/compare/@jobber/components@5.27.2...@jobber/components@5.28.0) (2024-08-27)
+
+
+### Features
+
+* **components:** allow custom title for Disclosure ([#1992](https://github.com/GetJobber/atlantis/issues/1992)) ([2b6e6cd](https://github.com/GetJobber/atlantis/commit/2b6e6cdb13a3afad9ca85ff163406cac674f1cbd))
+
+
+
+
+
 ## [5.27.2](https://github.com/GetJobber/atlantis/compare/@jobber/components@5.27.1...@jobber/components@5.27.2) (2024-08-26)